@@ -18,19 +18,19 @@
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.2.2",
 		"@sveltejs/adapter-vercel": "^5.3.2",
-		"@sveltejs/kit": "^2.5.11",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/bcrypt": "^5.0.2",
 		"@types/node": "^20.14.2",
 		"prettier": "^3.3.2",
-		"prettier-plugin-svelte": "^3.2.4",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prisma": "^5.16.1",
-		"svelte": "^4.2.18",
-		"svelte-check": "^3.8.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"ts-node": "^10.9.2",
 		"tslib": "^2.6.3",
-		"typescript": "^5.4.5",
-		"vite": "^5.3.0"
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4"
 	},
 	"dependencies": {
 		"@fortawesome/free-solid-svg-icons": "^6.5.2",
