@@ -1,4 +1,5 @@
 import { lucia } from "$lib/auth"
+import { error } from "@sveltejs/kit";
 
 export const handle = async ({ event, resolve }) => {
 	// get cookies from browser
@@ -26,5 +27,5 @@ export const handle = async ({ event, resolve }) => {
 	}
 	event.locals.user = user
 	event.locals.session = session
-	return resolve(event)
+	return await resolve(event)
 }
